WOR-1164 Mitigate deletion of virtualNetworkLink
