+++ service/src/main/java/bio/terra/landingzone/stairway/flight/create/resource/step/CreateVirtualNetworkLinkStep.java
@@ -1,5 +1,6 @@
 package bio.terra.landingzone.stairway.flight.create.resource.step;
 
+import bio.terra.landingzone.common.VirtualNetworkLinkResourceHelper;
 import bio.terra.landingzone.library.landingzones.definition.ArmManagers;
 import bio.terra.landingzone.library.landingzones.definition.ResourceNameGenerator;
 import bio.terra.landingzone.library.landingzones.definition.factories.ParametersResolver;
@@ -79,22 +80,7 @@ public class CreateVirtualNetworkLinkStep extends BaseResourceCreateStep {
 
   @Override
   protected void deleteResource(String resourceId) {
-    var idParts = resourceId.split("/");
-    if (idParts.length != 11) {
-      throw new IllegalArgumentException(
-          String.format("Invalid vnet link resourceId: %s", resourceId));
-    }
-    var mrgName = idParts[4];
-    var privateDnsZoneName = idParts[8];
-    var vnetLinkName = idParts[10];
-
-    armManagers
-        .azureResourceManager()
-        .privateDnsZones()
-        .manager()
-        .serviceClient()
-        .getVirtualNetworkLinks()
-        .delete(mrgName, privateDnsZoneName, vnetLinkName);
+    VirtualNetworkLinkResourceHelper.delete(armManagers, resourceId);
   }
 
   @Override
